temp: 원본 코드와 리팩토링 코드 비교용 커밋
