+++ src/components/intro/StudioToon.tsx
@@ -1,9 +1,9 @@
-import { useEffect, useState } from "react";
+import { useEffect, useState, useRef } from "react";
 
 import SectionTitle from "@components/commons/SectionTitle";
-import Modal from "@components/commons/Modal";
+import ToonModal from "./toon/ToonModal";
 
-import { useModal } from "@hooks/modal.hook";
+import { useModal } from "@hooks/new-modal.hook";
 import useIntersectionObserver from "@hooks/scroll.hook";
 
 import {
@@ -19,35 +19,45 @@ import {
 
 import type { FC } from "react";
 
+interface ToonItemData {
+  title: string;
+  pageCount: number;
+}
+
 const StudioToon: FC = () => {
   const IO_THRESHOLD = 0.3;
-
-  const TOON_ITEMS = [
+  const TOON_ITEMS: ToonItemData[] = [
     {
-      thumbnail: "/images/toon-thumb01.png",
       title: "DXI 개발하는 많화",
+      pageCount: 1,
     },
     {
-      thumbnail: "/images/toon-thumb02.png",
       title: "로키 디자인 회의",
+      pageCount: 3,
     },
     {
-      thumbnail: "/images/toon-thumb03.png",
       title: "스튜디오 함부기",
+      pageCount: 1,
     },
     {
-      thumbnail: "/images/toon-thumb04.png",
       title: "에다 처벌내규",
+      pageCount: 1,
     },
   ];
 
   const [childrenCount, setChildrenCount] = useState(0);
-  const [itemId, setItemId] = useState<number | null>(null);
-  const { isModalOpen, toggleModal } = useModal({
-    itemNumber: itemId || 0,
-    setItemId,
-  });
+  const [itemId, setItemId] = useState(0);
+
+  const modalRef = useRef<HTMLDivElement | null>(null);
+
+  const { displayModal, isModalOpen, toggleModal } = useModal(modalRef, 1500);
+
   const { targetRefs, entries } = useIntersectionObserver(IO_THRESHOLD);
+
+  const handleClickItem = (itemId: number) => (e: React.MouseEvent) => {
+    setItemId(itemId);
+    toggleModal(e);
+  };
 
   useEffect(() => {
     const toonItemWrapper = document.querySelector("#toon-item-wrapper");
@@ -57,32 +67,17 @@ const StudioToon: FC = () => {
     }
   }, []);
 
-  const handleItemClick = (id: number) => {
-    if (itemId === id) {
-      setItemId(null);
-      setTimeout(() => {
-        setItemId(id);
-      }, 0);
-    } else {
-      setItemId(id);
-    }
-  };
-
-  useEffect(() => {
-    if (itemId !== null) {
-      toggleModal();
-    }
-  }, [itemId]);
-
   return (
     <StudioToonBox
       ref={(el) => (targetRefs.current[0] = el as HTMLDivElement)}
       id="toon"
     >
-      <Modal
-        type="toon"
-        itemId={itemId || 0}
-        modalOpen={isModalOpen}
+      <ToonModal
+        itemId={itemId}
+        pages={TOON_ITEMS[itemId].pageCount}
+        modalRef={modalRef}
+        displayModal={displayModal}
+        isModalOpen={isModalOpen}
         toggleModal={toggleModal}
       />
 
@@ -111,10 +106,12 @@ const StudioToon: FC = () => {
           {TOON_ITEMS.map((toon, index) => (
             <ToonItem
               key={`EDDA Toon - ${index}`}
-              onClick={() => handleItemClick(index)}
+              onClick={handleClickItem(index)}
             >
               <img
-                src={toon.thumbnail}
+                src={`/images/toon-thumb${(index + 1)
+                  .toString()
+                  .padStart(2, "0")}.png`}
                 alt={`EDDA Toon - ${index} thumbnail`}
               />
               <ToonInfo>
@@ -127,7 +124,6 @@ const StudioToon: FC = () => {
           ))}
         </ToonItemWrapper>
       </StudioToonWrapper>
-
       <ToonVisual />
     </StudioToonBox>
   );
